fix(cli): preserve timestamp diff parity
